Move Key Connector migration check to unlock command
