+++ src/commands/login.command.ts
@@ -1,5 +1,4 @@
 import * as program from "commander";
-import * as inquirer from "inquirer";
 
 import { ApiService } from "jslib-common/abstractions/api.service";
 import { AuthService } from "jslib-common/abstractions/auth.service";
@@ -14,7 +13,6 @@ import { PolicyService } from "jslib-common/abstractions/policy.service";
 import { StateService } from "jslib-common/abstractions/state.service";
 import { SyncService } from "jslib-common/abstractions/sync.service";
 
-import { Response } from "jslib-node/cli/models/response";
 import { MessageResponse } from "jslib-node/cli/models/response/messageResponse";
 
 import { Utils } from "jslib-common/misc/utils";
@@ -28,6 +26,7 @@ export class LoginCommand extends BaseLoginCommand {
     authService: AuthService,
     apiService: ApiService,
     cryptoFunctionService: CryptoFunctionService,
+    syncService: SyncService,
     i18nService: I18nService,
     environmentService: EnvironmentService,
     passwordGenerationService: PasswordGenerationService,
@@ -35,7 +34,6 @@ export class LoginCommand extends BaseLoginCommand {
     stateService: StateService,
     cryptoService: CryptoService,
     policyService: PolicyService,
-    private syncService: SyncService,
     private keyConnectorService: KeyConnectorService,
     private logoutCallback: () => Promise<void>
   ) {
@@ -50,12 +48,47 @@ export class LoginCommand extends BaseLoginCommand {
       stateService,
       cryptoService,
       policyService,
-      "cli"
+      "cli",
+      syncService
     );
     this.logout = this.logoutCallback;
     this.validatedParams = async () => {
       const key = await cryptoFunctionService.randomBytes(64);
       process.env.BW_SESSION = Utils.fromBufferToB64(key);
+    };
+    this.success = async () => {
+      await this.syncService.fullSync(true);
+
+      const usesKeyConnector = await this.keyConnectorService.getUsesKeyConnector();
+
+      if (
+        (this.options.sso != null || this.options.apikey != null) &&
+        this.canInteract &&
+        !usesKeyConnector
+      ) {
+        const res = new MessageResponse(
+          "You are logged in!",
+          "\n" + "To unlock your vault, use the `unlock` command. ex:\n" + "$ bw unlock"
+        );
+        return res;
+      } else {
+        const res = new MessageResponse(
+          "You are logged in!",
+          "\n" +
+            "To unlock your vault, set your session key to the `BW_SESSION` environment variable. ex:\n" +
+            '$ export BW_SESSION="' +
+            process.env.BW_SESSION +
+            '"\n' +
+            '> $env:BW_SESSION="' +
+            process.env.BW_SESSION +
+            '"\n\n' +
+            "You can also pass the session key to any command with the `--session` option. ex:\n" +
+            "$ bw list items --session " +
+            process.env.BW_SESSION
+        );
+        res.raw = process.env.BW_SESSION;
+        return res;
+      }
     };
   }
 
@@ -63,105 +96,5 @@ export class LoginCommand extends BaseLoginCommand {
     this.options = options;
     this.email = email;
     return super.run(email, password, options);
-  }
-
-  async onSuccessfulLogin(): Promise<Response> {
-    // Full sync required for key connector check
-    await this.syncService.fullSync(true);
-
-    if (await this.keyConnectorService.userNeedsMigration()) {
-      return await this.migrateToKeyConnector();
-    }
-
-    return this.successResponse();
-  }
-
-  private async successResponse() {
-    const usesKeyConnector = await this.keyConnectorService.getUsesKeyConnector();
-
-    if (
-      (this.options.sso != null || this.options.apikey != null) &&
-      this.canInteract &&
-      !usesKeyConnector
-    ) {
-      const message = new MessageResponse(
-        "You are logged in!",
-        "\n" + "To unlock your vault, use the `unlock` command. ex:\n" + "$ bw unlock"
-      );
-      return Response.success(message);
-    } else {
-      const message = new MessageResponse(
-        "You are logged in!",
-        "\n" +
-          "To unlock your vault, set your session key to the `BW_SESSION` environment variable. ex:\n" +
-          '$ export BW_SESSION="' +
-          process.env.BW_SESSION +
-          '"\n' +
-          '> $env:BW_SESSION="' +
-          process.env.BW_SESSION +
-          '"\n\n' +
-          "You can also pass the session key to any command with the `--session` option. ex:\n" +
-          "$ bw list items --session " +
-          process.env.BW_SESSION
-      );
-      message.raw = process.env.BW_SESSION;
-      return Response.success(message);
-    }
-  }
-
-  private async migrateToKeyConnector(): Promise<Response> {
-    // If no interaction available, alert user to use web vault
-    if (!this.canInteract) {
-      await this.logout();
-      return Response.error(
-        new MessageResponse(
-          "An organization you are a member of is using Key Connector. " +
-            "In order to access the vault, you must opt-in to Key Connector now via the web vault. You have been logged out.",
-          null
-        )
-      );
-    }
-
-    const organization = await this.keyConnectorService.getManagingOrganization();
-
-    const answer: inquirer.Answers = await inquirer.createPromptModule({ output: process.stderr })({
-      type: "list",
-      name: "convert",
-      message:
-        organization.name +
-        " is using a self-hosted key server. A master password is no longer required to log in for members of this organization. ",
-      choices: [
-        {
-          name: "Remove master password and log in",
-          value: "remove",
-        },
-        {
-          name: "Leave organization and log in",
-          value: "leave",
-        },
-        {
-          name: "Exit",
-          value: "exit",
-        },
-      ],
-    });
-
-    if (answer.convert === "remove") {
-      await this.keyConnectorService.migrateUser();
-
-      // Update environment URL - required for api key login
-      const urls = this.environmentService.getUrls();
-      urls.keyConnector = organization.keyConnectorUrl;
-      await this.environmentService.setUrls(urls, true);
-
-      return this.successResponse();
-    } else if (answer.convert === "leave") {
-      await this.apiService.postLeaveOrganization(organization.id);
-      await this.syncService.fullSync(true);
-      return this.successResponse();
-    } else {
-      await this.logout();
-      return Response.error("You have been logged out.");
-    }
   }
 }
