+++ src/commands/unlock.command.ts
@@ -4,7 +4,11 @@ import * as inquirer from "inquirer";
 import { ApiService } from "jslib-common/abstractions/api.service";
 import { CryptoService } from "jslib-common/abstractions/crypto.service";
 import { CryptoFunctionService } from "jslib-common/abstractions/cryptoFunction.service";
+import { EnvironmentService } from "jslib-common/abstractions/environment.service";
+import { KeyConnectorService } from "jslib-common/abstractions/keyConnector.service";
 import { StateService } from "jslib-common/abstractions/state.service";
+import { SyncService } from "jslib-common/abstractions/sync.service";
+import { ConsoleLogService } from "jslib-common/services/consoleLog.service";
 
 import { Response } from "jslib-node/cli/models/response";
 import { MessageResponse } from "jslib-node/cli/models/response/messageResponse";
@@ -15,7 +19,6 @@ import { Utils } from "jslib-common/misc/utils";
 
 import { HashPurpose } from "jslib-common/enums/hashPurpose";
 import { NodeUtils } from "jslib-common/misc/nodeUtils";
-import { ConsoleLogService } from "jslib-common/services/consoleLog.service";
 
 export class UnlockCommand {
   constructor(
@@ -23,7 +26,11 @@ export class UnlockCommand {
     private stateService: StateService,
     private cryptoFunctionService: CryptoFunctionService,
     private apiService: ApiService,
-    private logService: ConsoleLogService
+    private logService: ConsoleLogService,
+    private keyConnectorService: KeyConnectorService,
+    private environmentService: EnvironmentService,
+    private syncService: SyncService,
+    private logout: () => Promise<void>
   ) {}
 
   async run(password: string, options: program.OptionValues) {
@@ -92,22 +99,12 @@ export class UnlockCommand {
 
     if (passwordValid) {
       await this.cryptoService.setKey(key);
-      const res = new MessageResponse(
-        "Your vault is now unlocked!",
-        "\n" +
-          "To unlock your vault, set your session key to the `BW_SESSION` environment variable. ex:\n" +
-          '$ export BW_SESSION="' +
-          process.env.BW_SESSION +
-          '"\n' +
-          '> $env:BW_SESSION="' +
-          process.env.BW_SESSION +
-          '"\n\n' +
-          "You can also pass the session key to any command with the `--session` option. ex:\n" +
-          "$ bw list items --session " +
-          process.env.BW_SESSION
-      );
-      res.raw = process.env.BW_SESSION;
-      return Response.success(res);
+
+      if (await this.keyConnectorService.userNeedsMigration()) {
+        return await this.migrateToKeyConnector();
+      }
+
+      return this.successResponse();
     } else {
       return Response.error("Invalid master password.");
     }
@@ -116,5 +113,86 @@ export class UnlockCommand {
   private async setNewSessionKey() {
     const key = await this.cryptoFunctionService.randomBytes(64);
     process.env.BW_SESSION = Utils.fromBufferToB64(key);
+  }
+
+  private async successResponse() {
+    const res = new MessageResponse(
+      "Your vault is now unlocked!",
+      "\n" +
+        "To unlock your vault, set your session key to the `BW_SESSION` environment variable. ex:\n" +
+        '$ export BW_SESSION="' +
+        process.env.BW_SESSION +
+        '"\n' +
+        '> $env:BW_SESSION="' +
+        process.env.BW_SESSION +
+        '"\n\n' +
+        "You can also pass the session key to any command with the `--session` option. ex:\n" +
+        "$ bw list items --session " +
+        process.env.BW_SESSION
+    );
+    res.raw = process.env.BW_SESSION;
+    return Response.success(res);
+  }
+
+  private async migrateToKeyConnector(): Promise<Response> {
+    // If no interaction available, alert user to use web vault
+    const canInteract = process.env.BW_NOINTERACTION !== "true";
+    if (!canInteract) {
+      await this.logout();
+      return Response.error(
+        new MessageResponse(
+          "An organization you are a member of is using Key Connector. " +
+            "In order to access the vault, you must opt-in to Key Connector now via the web vault. You have been logged out.",
+          null
+        )
+      );
+    }
+
+    const organization = await this.keyConnectorService.getManagingOrganization();
+
+    const answer: inquirer.Answers = await inquirer.createPromptModule({ output: process.stderr })({
+      type: "list",
+      name: "convert",
+      message:
+        organization.name +
+        " is using a self-hosted key server. A master password is no longer required to log in for members of this organization. ",
+      choices: [
+        {
+          name: "Remove master password and unlock",
+          value: "remove",
+        },
+        {
+          name: "Leave organization and unlock",
+          value: "leave",
+        },
+        {
+          name: "Log out",
+          value: "exit",
+        },
+      ],
+    });
+
+    if (answer.convert === "remove") {
+      try {
+        await this.keyConnectorService.migrateUser();
+      } catch (e) {
+        await this.logout();
+        throw e;
+      }
+
+      // Update environment URL - required for api key login
+      const urls = this.environmentService.getUrls();
+      urls.keyConnector = organization.keyConnectorUrl;
+      await this.environmentService.setUrls(urls, true);
+
+      return this.successResponse();
+    } else if (answer.convert === "leave") {
+      await this.apiService.postLeaveOrganization(organization.id);
+      await this.syncService.fullSync(true);
+      return this.successResponse();
+    } else {
+      await this.logout();
+      return Response.error("You have been logged out.");
+    }
   }
 }
