+++ src/program.ts
@@ -148,6 +148,7 @@ export class Program extends BaseProgram {
             this.main.authService,
             this.main.apiService,
             this.main.cryptoFunctionService,
+            this.main.syncService,
             this.main.i18nService,
             this.main.environmentService,
             this.main.passwordGenerationService,
@@ -155,7 +156,6 @@ export class Program extends BaseProgram {
             this.main.stateService,
             this.main.cryptoService,
             this.main.policyService,
-            this.main.syncService,
             this.main.keyConnectorService,
             async () => await this.main.logout()
           );
@@ -256,7 +256,11 @@ export class Program extends BaseProgram {
             this.main.stateService,
             this.main.cryptoFunctionService,
             this.main.apiService,
-            this.main.logService
+            this.main.logService,
+            this.main.keyConnectorService,
+            this.main.environmentService,
+            this.main.syncService,
+            this.main.logout
           );
           const response = await command.run(password, cmd);
           this.processResponse(response);
@@ -494,7 +498,11 @@ export class Program extends BaseProgram {
           this.main.stateService,
           this.main.cryptoFunctionService,
           this.main.apiService,
-          this.main.logService
+          this.main.logService,
+          this.main.keyConnectorService,
+          this.main.environmentService,
+          this.main.syncService,
+          this.main.logout
         );
         const response = await command.run(null, null);
         if (!response.success) {
